alertsystem_service/requirements.txt
mysql-connector-python
confluent-kafka